generalized items.router to any single-list API of vcurrency (items, packs, etc.)
