+++ server/vcurrency/index.js
@@ -1,9 +1,16 @@
 'use strict';
 
 const express = require('express');
+const proxy = require('./single-list-proxy');
+const config = require('../../config');
 
 const router = new express.Router();
 
-router.use('/items', require('./items.router'));
+router.use('/items', proxy('/products', (json) => {
+  return {
+    items: json,
+    currencies: config.services.virtualcurrency.currencies
+  };
+}));
 
 module.exports = router;
